requirements.txt
flask
python-pptx
beautifulsoup4
